+++ packages/app/app/composables/message/friend/useReadFriends.ts
@@ -1,16 +1,16 @@
-import { useFriendStore } from "@/store/message/user/friend";
-
 export const useReadFriends = () => {
   const { $trpc } = useNuxtApp();
   const friendStore = useFriendStore();
-  const { blockedUsers, friends, pendingRequests, sentRequests } = storeToRefs(friendStore);
+  const friendRequestStore = useFriendRequestStore();
+  const { blockedUsers, friends } = storeToRefs(friendStore);
+  const { pendingRequests, sentRequests } = storeToRefs(friendRequestStore);
 
   const readFriends = async () => {
     const [readBlockedUsers, readFriends, readPendingRequests, readSentRequests] = await Promise.all([
       $trpc.friend.readBlockedUsers.query(),
       $trpc.friend.readFriends.query(),
-      $trpc.friend.readPendingRequests.query(),
-      $trpc.friend.readSentRequests.query(),
+      $trpc.friendRequest.readPendingRequests.query(),
+      $trpc.friendRequest.readSentRequests.query(),
     ]);
     blockedUsers.value = readBlockedUsers;
     friends.value = readFriends;
