+++ packages/app/app/composables/message/subscribables/useFriendSubscribables.ts
@@ -1,29 +1,29 @@
 import { authClient } from "@/services/auth/authClient";
-import { useFriendStore } from "@/store/message/user/friend";
 
 export const useFriendSubscribables = () => {
   const session = authClient.useSession();
   const { $trpc } = useNuxtApp();
   const friendStore = useFriendStore();
-  const { storeAcceptFriendRequest, storeCreatePendingRequest, storeDeclineFriendRequest, storeDeleteFriend } =
-    friendStore;
+  const friendRequestStore = useFriendRequestStore();
+  const { storeDeleteFriend } = friendStore;
+  const { storeAcceptFriendRequest, storeCreatePendingRequest, storeDeclineFriendRequest } = friendRequestStore;
 
   useOnlineSubscribable(
     () => session.value.data?.user.id,
     (userId) => {
       if (!userId) return undefined;
 
-      const sendFriendRequestUnsubscribable = $trpc.friend.onSendFriendRequest.subscribe(undefined, {
+      const sendFriendRequestUnsubscribable = $trpc.friendRequest.onSendFriendRequest.subscribe(undefined, {
         onData: (senderUser) => {
           storeCreatePendingRequest(senderUser);
         },
       });
-      const acceptFriendRequestUnsubscribable = $trpc.friend.onAcceptFriendRequest.subscribe(undefined, {
+      const acceptFriendRequestUnsubscribable = $trpc.friendRequest.onAcceptFriendRequest.subscribe(undefined, {
         onData: (receiverUser) => {
           storeAcceptFriendRequest(receiverUser);
         },
       });
-      const declineFriendRequestUnsubscribable = $trpc.friend.onDeclineFriendRequest.subscribe(undefined, {
+      const declineFriendRequestUnsubscribable = $trpc.friendRequest.onDeclineFriendRequest.subscribe(undefined, {
         onData: (declinerId) => {
           storeDeclineFriendRequest(declinerId);
         },
