+++ packages/app/server/trpc/routers/friend.ts
@@ -1,51 +1,23 @@
-import type { FriendRequestNotificationEventGridData, User } from "@esposter/db-schema";
+import type { User } from "@esposter/db-schema";
 
 import { friendUserIdInputSchema } from "#shared/models/db/friend/FriendUserIdInput";
 import { searchUsersInputSchema } from "#shared/models/db/friend/SearchUsersInput";
 import { MAX_READ_LIMIT } from "#shared/services/pagination/constants";
-import { useEventGridPublisherClient } from "@@/server/composables/azure/eventGrid/useEventGridPublisherClient";
 import { escapeLike } from "@@/server/services/db/escapeLike";
 import { on } from "@@/server/services/events/on";
 import { getFriendshipId } from "@@/server/services/friend/getFriendshipId";
 import { friendEventEmitter } from "@@/server/services/message/events/friendEventEmitter";
 import { router } from "@@/server/trpc";
 import { standardAuthedProcedure } from "@@/server/trpc/procedure/standardAuthedProcedure";
-import { getPushSubscriptionsForUser } from "@esposter/db";
-import { AzureFunction, blocks, DatabaseEntityType, friendRequests, friends, users } from "@esposter/db-schema";
+import { blocks, DatabaseEntityType, friendRequests, friends, users } from "@esposter/db-schema";
 import { InvalidOperationError, Operation } from "@esposter/shared";
 import { TRPCError } from "@trpc/server";
-import { and, eq, getTableColumns, isNull, ilike, ne, or } from "drizzle-orm";
+import { and, eq, getTableColumns, ilike, isNull, ne, or } from "drizzle-orm";
 
 export type { FriendUserIdInput } from "#shared/models/db/friend/FriendUserIdInput";
 export type { SearchUsersInput } from "#shared/models/db/friend/SearchUsersInput";
 
 export const friendRouter = router({
-  acceptFriendRequest: standardAuthedProcedure
-    .input(friendUserIdInputSchema)
-    .mutation(async ({ ctx, input: senderId }) => {
-      const userId = ctx.getSessionPayload.user.id;
-      const friendshipId = getFriendshipId(senderId, userId);
-      const [acceptedFriend] = await ctx.db.transaction(async (tx) => {
-        const [deletedRequest] = await tx
-          .delete(friendRequests)
-          .where(and(eq(friendRequests.id, friendshipId), eq(friendRequests.receiverId, userId)))
-          .returning();
-        if (!deletedRequest) return [];
-        return tx.insert(friends).values({ id: friendshipId, receiverId: userId, senderId }).returning();
-      });
-      if (!acceptedFriend)
-        throw new TRPCError({
-          code: "NOT_FOUND",
-          message: new InvalidOperationError(Operation.Update, DatabaseEntityType.Friend, friendshipId).message,
-        });
-      const receiverUser: User = {
-        ...ctx.getSessionPayload.user,
-        deletedAt: null,
-        image: ctx.getSessionPayload.user.image ?? null,
-      };
-      friendEventEmitter.emit("acceptFriendRequest", { receiverUser, senderId });
-      return acceptedFriend;
-    }),
   blockUser: standardAuthedProcedure.input(friendUserIdInputSchema).mutation(async ({ ctx, input: targetUserId }) => {
     const userId = ctx.getSessionPayload.user.id;
     if (userId === targetUserId)
@@ -57,7 +29,9 @@ export const friendRouter = router({
     await ctx.db.delete(friends).where(eq(friends.id, friendshipId));
     await ctx.db.delete(friendRequests).where(eq(friendRequests.id, friendshipId));
     await ctx.db.insert(blocks).values({ blockedId: targetUserId, blockerId: userId }).onConflictDoNothing();
-    const blockedUser = await ctx.db.query.users.findFirst({ where: eq(users.id, targetUserId) });
+    const blockedUser = await ctx.db.query.users.findFirst({
+      where: (users, { eq }) => eq(users.id, targetUserId),
+    });
     if (!blockedUser)
       throw new TRPCError({
         code: "NOT_FOUND",
@@ -65,36 +39,11 @@ export const friendRouter = router({
       });
     return blockedUser;
   }),
-  declineFriendRequest: standardAuthedProcedure
-    .input(friendUserIdInputSchema)
-    .mutation(async ({ ctx, input: senderId }) => {
-      const userId = ctx.getSessionPayload.user.id;
-      const friendshipId = getFriendshipId(senderId, userId);
-      const [declinedRequest] = await ctx.db
-        .delete(friendRequests)
-        .where(and(eq(friendRequests.id, friendshipId), eq(friendRequests.receiverId, userId)))
-        .returning();
-      if (declinedRequest) friendEventEmitter.emit("declineFriendRequest", { receiverId: userId, senderId });
-    }),
   deleteFriend: standardAuthedProcedure.input(friendUserIdInputSchema).mutation(async ({ ctx, input: friendId }) => {
     const userId = ctx.getSessionPayload.user.id;
     const friendshipId = getFriendshipId(userId, friendId);
     const [deletedFriend] = await ctx.db.delete(friends).where(eq(friends.id, friendshipId)).returning();
     if (deletedFriend) friendEventEmitter.emit("deleteFriend", { receiverId: friendId, senderId: userId });
-  }),
-  onAcceptFriendRequest: standardAuthedProcedure.subscription(async function* ({ ctx, signal }) {
-    const userId = ctx.getSessionPayload.user.id;
-    for await (const [{ receiverUser, senderId }] of on(friendEventEmitter, "acceptFriendRequest", { signal })) {
-      if (senderId !== userId) continue;
-      yield receiverUser;
-    }
-  }),
-  onDeclineFriendRequest: standardAuthedProcedure.subscription(async function* ({ ctx, signal }) {
-    const userId = ctx.getSessionPayload.user.id;
-    for await (const [{ receiverId, senderId }] of on(friendEventEmitter, "declineFriendRequest", { signal })) {
-      if (senderId !== userId) continue;
-      yield receiverId;
-    }
   }),
   onDeleteFriend: standardAuthedProcedure.subscription(async function* ({ ctx, signal }) {
     const userId = ctx.getSessionPayload.user.id;
@@ -103,17 +52,10 @@ export const friendRouter = router({
       yield senderId;
     }
   }),
-  onSendFriendRequest: standardAuthedProcedure.subscription(async function* ({ ctx, signal }) {
-    const userId = ctx.getSessionPayload.user.id;
-    for await (const [{ receiverId, senderUser }] of on(friendEventEmitter, "sendFriendRequest", { signal })) {
-      if (receiverId !== userId) continue;
-      yield senderUser;
-    }
-  }),
   readBlockedUsers: standardAuthedProcedure.query<User[]>(async ({ ctx }) => {
     const userId = ctx.getSessionPayload.user.id;
     const blockedRows = await ctx.db.query.blocks.findMany({
-      where: (b, { eq }) => eq(b.blockerId, userId),
+      where: (blocks, { eq }) => eq(blocks.blockerId, userId),
       with: { blocked: true },
     });
     return blockedRows.map(({ blocked }) => blocked);
@@ -131,22 +73,6 @@ export const friendRouter = router({
         ),
       );
   }),
-  readPendingRequests: standardAuthedProcedure.query<User[]>(async ({ ctx }) => {
-    const userId = ctx.getSessionPayload.user.id;
-    const pendingRequests = await ctx.db.query.friendRequests.findMany({
-      where: (fr, { eq }) => eq(fr.receiverId, userId),
-      with: { sender: true },
-    });
-    return pendingRequests.map(({ sender }) => sender);
-  }),
-  readSentRequests: standardAuthedProcedure.query<User[]>(async ({ ctx }) => {
-    const userId = ctx.getSessionPayload.user.id;
-    const sentRequests = await ctx.db.query.friendRequests.findMany({
-      where: (fr, { eq }) => eq(fr.senderId, userId),
-      with: { receiver: true },
-    });
-    return sentRequests.map(({ receiver }) => receiver);
-  }),
   searchUsers: standardAuthedProcedure.input(searchUsersInputSchema).query(({ ctx, input: name }) => {
     const userId = ctx.getSessionPayload.user.id;
     const blockedSubquery = ctx.db
@@ -162,73 +88,6 @@ export const friendRouter = router({
       .where(and(ilike(users.name, `%${escapeLike(name)}%`), ne(users.id, userId), isNull(blockedSubquery.id)))
       .limit(MAX_READ_LIMIT);
   }),
-  sendFriendRequest: standardAuthedProcedure
-    .input(friendUserIdInputSchema)
-    .mutation(async ({ ctx, input: receiverId }) => {
-      const userId = ctx.getSessionPayload.user.id;
-      if (userId === receiverId)
-        throw new TRPCError({
-          code: "BAD_REQUEST",
-          message: new InvalidOperationError(Operation.Create, DatabaseEntityType.Friend, userId).message,
-        });
-      const existingBlock = await ctx.db.query.blocks.findFirst({
-        where: (b, { and, or, eq }) =>
-          or(
-            and(eq(b.blockerId, userId), eq(b.blockedId, receiverId)),
-            and(eq(b.blockerId, receiverId), eq(b.blockedId, userId)),
-          ),
-      });
-      if (existingBlock)
-        throw new TRPCError({
-          code: "BAD_REQUEST",
-          message: new InvalidOperationError(Operation.Create, DatabaseEntityType.Friend, receiverId).message,
-        });
-      const friendshipId = getFriendshipId(userId, receiverId);
-      const [newRequest] = await ctx.db
-        .insert(friendRequests)
-        .values({ id: friendshipId, receiverId, senderId: userId })
-        .onConflictDoNothing({ target: friendRequests.id })
-        .returning();
-      const request =
-        newRequest ?? (await ctx.db.query.friendRequests.findFirst({ where: eq(friendRequests.id, friendshipId) }));
-      if (!request)
-        throw new TRPCError({
-          code: "BAD_REQUEST",
-          message: new InvalidOperationError(Operation.Create, DatabaseEntityType.FriendRequest, friendshipId).message,
-        });
-      if (newRequest) {
-        const senderUser: User = {
-          ...ctx.getSessionPayload.user,
-          deletedAt: null,
-          image: ctx.getSessionPayload.user.image ?? null,
-        };
-        friendEventEmitter.emit("sendFriendRequest", { receiverId, senderUser });
-
-        const readPushSubscriptions = await getPushSubscriptionsForUser(ctx.db, receiverId);
-        if (readPushSubscriptions.length > 0) {
-          const eventGridPublisherClient = useEventGridPublisherClient();
-          const data: FriendRequestNotificationEventGridData = {
-            notificationOptions: { icon: senderUser.image, title: senderUser.name },
-            receiverId,
-          };
-          await eventGridPublisherClient.send([
-            {
-              data,
-              dataVersion: "1.0",
-              eventType: AzureFunction.ProcessFriendRequestNotification,
-              subject: `${userId}/${receiverId}`,
-            },
-          ]);
-        }
-      }
-      const receiverUser = await ctx.db.query.users.findFirst({ where: eq(users.id, receiverId) });
-      if (!receiverUser)
-        throw new TRPCError({
-          code: "NOT_FOUND",
-          message: new InvalidOperationError(Operation.Read, DatabaseEntityType.Friend, receiverId).message,
-        });
-      return receiverUser;
-    }),
   unblockUser: standardAuthedProcedure
     .input(friendUserIdInputSchema)
     .mutation(async ({ ctx, input: blockedUserId }) => {
