+++ packages/db-schema/src/schema.ts
@@ -2,6 +2,7 @@ import * as accounts from "@/schema/accounts";
 import * as achievements from "@/schema/achievements";
 import * as appUsers from "@/schema/appUsers";
 import * as blocks from "@/schema/blocks";
+import * as friendRequests from "@/schema/friendRequests";
 import * as friends from "@/schema/friends";
 import * as invites from "@/schema/invites";
 import * as likes from "@/schema/likes";
@@ -24,6 +25,7 @@ export const schema = {
   ...achievements,
   ...appUsers,
   ...blocks,
+  ...friendRequests,
   ...friends,
   ...invites,
   ...likes,
